Batching and serial number fixes
Updates are now batched (or at least batchable), meaning that we don't
need to read, parse, edit, and save zonefiles N times for N changes.
This drops the time needed for 900 changes from ~10s down to ~3s.
Since 'diff' takes 2.5s, that's probably about as fast as it's going to
get on this config.

Serial numbers are now updated for zonefiles.  The code handles the two
common cases (simple incrementing integer and YYYYMMDDxx) as expected.

Signed-off-by: Scott Laird <scott@sigkill.org>